requirements.txt
numpy~=1.23.2
opencv-python==4.6.0.66
pybind11==2.10.0
imutils~=0.5.4
Pillow~=7.0.0